Revert "extends internally used digitizerID to be unique"
This reverts commit aeee35675fc1f4047cd173c3b4f32ae34e7a36ba.
Width of 64bits not strictly necessary and implementation in EFU currently expects 32bits only.

+++ DataFormat.hpp
@@ -60,7 +60,7 @@ namespace Data
     {
         uint64_t runID;
         uint64_t globalTime;
-        uint64_t digitizerID;
+        uint32_t digitizerID;
         uint16_t elementType;
         uint16_t numElements;
         uint16_t version;
